requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
pkg-resources==0.0.0
psycopg2==2.7.6
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
